requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
streamlit>=1.28.0
plotly>=5.17.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0
PyYAML>=6.0
python-dotenv>=1.0.0
rich>=13.6.0
click>=8.1.0
pyarrow>=13.0.0
openpyxl>=3.1.0
pytest>=7.4.0
httpx>=0.25.0
